+++ source/QuantizerModule.cpp
@@ -79,7 +79,7 @@ void QuantizerModule::processSample(double& sample) {
 // Method 2: Without bit shifting
 //------------------------------------
 	// Convert signed to unsigned
-	long unsigned tu = temp + min32bitValueSigned;
+	long unsigned tu = temp + (long unsigned)min32bitValueSigned;
 
 	// Those variables are fixed for the specified bitsToReset! => needs to be done only one time!
 	unsigned long newMax = (pow(2, 32 - bitsToReset) - 1) * (pow(2, bitsToReset) - 1);
